requirements.txt
docopt==0.6.2
python-jenkins==1.3.0
